[IMP] base_report_to_printer - split code in multiple files
