+++ base_report_to_printer/AUTHORS.txt
@@ -2,3 +2,4 @@ Ferran Pegueroles <ferran@pegueroles.com>
 Albert Cervera i Areny <albert@nan-tic.com>
 Davide Corio <davide.corio@agilebg.com>
 Lorenzo Battistini <lorenzo.battistini@agilebg.com>
+Yannick Vaucher <yannick.vaucher@camptocamp.com>
