+++ base_report_to_printer/report_xml_action.py
@@ -22,9 +22,29 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-from . import printing
-from . import report_xml_action
-from . import report_service
-from . import users
-from . import ir_report
-from . import wizard
+from openerp.osv import orm, fields
+
+from printing import _available_action_types
+
+class report_xml_action(orm.Model):
+    _name = 'printing.report.xml.action'
+    _description = 'Report Printing Actions'
+    _columns = {
+        'report_id': fields.many2one('ir.actions.report.xml', 'Report', required=True, ondelete='cascade'),
+        'user_id': fields.many2one('res.users', 'User', required=True, ondelete='cascade'),
+        'action': fields.selection(_available_action_types, 'Action', required=True),
+        'printer_id': fields.many2one('printing.printer', 'Printer'),
+        }
+
+
+    def behaviour(self, cr, uid, act_id, context=None):
+        result = {}
+        if not act_id:
+            return False
+        action = self.browse(cr, uid, act_id, context=context)
+        return {
+            'action': action.action,
+            'printer': action.printer_id,
+            }
+
+# vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
